Remove old sampling method
