+++ tree_labeller/tree/selectors/__init__.py
@@ -1,3 +1,2 @@
 from .distant_leaves import select_distant_leaves
-from .sampling import sample
 from .top_down import select_top_down
